requirements.txt
django
factory-boy
pillow
PyJWT
django-allauth
cryptography
requests